requirements.txt
Django==5.2
Pillow==11.2.1
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2
